+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/domain/base/service/AbstractAssetBaseService.java
@@ -25,24 +25,28 @@ import org.eclipse.tractusx.traceability.assets.domain.base.IrsRepository;
 import org.eclipse.tractusx.traceability.assets.domain.base.model.AssetBase;
 import org.eclipse.tractusx.traceability.assets.domain.base.model.Owner;
 import org.eclipse.tractusx.traceability.assets.domain.base.model.QualityType;
+import org.eclipse.tractusx.traceability.assets.domain.base.model.SemanticDataModel;
 import org.eclipse.tractusx.traceability.assets.infrastructure.asbuilt.model.ManufacturingInfo;
-import org.eclipse.tractusx.traceability.common.model.SearchCriteria;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request.BomLifecycle;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.Direction;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.relationship.Aspect;
 import org.eclipse.tractusx.traceability.common.config.AssetsAsyncConfig;
 import org.eclipse.tractusx.traceability.common.model.PageResult;
+import org.eclipse.tractusx.traceability.common.model.SearchCriteria;
 import org.eclipse.tractusx.traceability.qualitynotification.domain.base.model.QualityNotification;
 import org.eclipse.tractusx.traceability.qualitynotification.domain.base.model.QualityNotificationStatus;
 import org.springframework.data.domain.Pageable;
 import org.springframework.scheduling.annotation.Async;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 @Slf4j
 public abstract class AbstractAssetBaseService implements AssetBaseService {
+
+    private static List<String> SUPPORTED_ENUM_FIELDS = List.of("owner", "qualityType", "semanticDataModel");
 
     protected abstract AssetRepository getAssetRepository();
 
@@ -54,7 +58,7 @@ public abstract class AbstractAssetBaseService implements AssetBaseService {
 
     protected abstract BomLifecycle getBomLifecycle();
 
-
+    @Override
     @Async(value = AssetsAsyncConfig.SYNCHRONIZE_ASSETS_EXECUTOR)
     public void synchronizeAssetsAsync(String globalAssetId) {
         log.info("Synchronizing assets for globalAssetId: {}", globalAssetId);
@@ -83,6 +87,7 @@ public abstract class AbstractAssetBaseService implements AssetBaseService {
         }
     }
 
+    @Override
     @Async(value = AssetsAsyncConfig.SYNCHRONIZE_ASSETS_EXECUTOR)
     public void synchronizeAssetsAsync(List<String> globalAssetIds) {
         for (String globalAssetId : globalAssetIds) {
@@ -94,7 +99,7 @@ public abstract class AbstractAssetBaseService implements AssetBaseService {
         }
     }
 
-
+    @Override
     public void setAssetsInvestigationStatus(QualityNotification investigation) {
         getAssetRepository().getAssetsById(investigation.getAssetIds()).forEach(asset -> {
             // Assets in status closed will be false, others true
@@ -103,6 +108,7 @@ public abstract class AbstractAssetBaseService implements AssetBaseService {
         });
     }
 
+    @Override
     public void setAssetsAlertStatus(QualityNotification alert) {
         getAssetRepository().getAssetsById(alert.getAssetIds()).forEach(asset -> {
             // Assets in status closed will be false, others true
@@ -111,8 +117,7 @@ public abstract class AbstractAssetBaseService implements AssetBaseService {
         });
     }
 
-
-
+    @Override
     public AssetBase updateQualityType(String assetId, QualityType qualityType) {
         AssetBase foundAsset = getAssetRepository().getAssetById(assetId);
         foundAsset.setQualityType(qualityType);
@@ -124,14 +129,17 @@ public abstract class AbstractAssetBaseService implements AssetBaseService {
         return getAssetRepository().getAssets(pageable, filter);
     }
 
+    @Override
     public AssetBase getAssetById(String assetId) {
         return getAssetRepository().getAssetById(assetId);
     }
 
+    @Override
     public List<AssetBase> getAssetsById(List<String> assetIds) {
         return getAssetRepository().getAssetsById(assetIds);
     }
 
+    @Override
     public AssetBase getAssetByChildId(String assetId, String childId) {
         return getAssetRepository().getAssetByChildId(assetId, childId);
     }
@@ -141,5 +149,26 @@ public abstract class AbstractAssetBaseService implements AssetBaseService {
         return getAssetRepository().getAssets().stream()
                 .collect(Collectors.groupingBy(
                         asset -> ManufacturingInfo.from(asset.getDetailAspectModels()).getManufacturingCountry(), Collectors.counting()));
+    }
+
+    @Override
+    public List<String> getDistinctFilterValues(String fieldName, Long size) {
+        if (isSupportedEnumType(fieldName)) {
+            return getAssetEnumFieldValues(fieldName);
+        }
+        return getAssetRepository().getFieldValues(fieldName, size);
+    }
+
+    private boolean isSupportedEnumType(String fieldName) {
+        return SUPPORTED_ENUM_FIELDS.contains(fieldName);
+    }
+
+    private List<String> getAssetEnumFieldValues(String fieldName) {
+        return switch (fieldName) {
+            case "owner" -> Arrays.stream(Owner.values()).map(Enum::name).toList();
+            case "qualityType" -> Arrays.stream(QualityType.values()).map(Enum::name).toList();
+            case "semanticDataModel" -> Arrays.stream(SemanticDataModel.values()).map(Enum::name).toList();
+            default -> null;
+        };
     }
 }