+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/IntegrationTestSpecification.java
@@ -78,9 +78,9 @@ public class IntegrationTestSpecification {
 
     protected void eventually(Callable<Boolean> conditions) throws InterruptedException {
         Awaitility.setDefaultPollInterval(500, TimeUnit.MILLISECONDS);
-        Awaitility.setDefaultTimeout(15, TimeUnit.SECONDS);
+        Awaitility.setDefaultTimeout(30, TimeUnit.SECONDS);
         Awaitility.pollInSameThread();
-        await().until(conditions, Matchers.equalTo(true));
+        await().pollDelay(2, TimeUnit.SECONDS).until(conditions, Matchers.equalTo(true));
     }
 
 }