+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - NEW API DELETE /api/registry
 - NEW API GET /api/shelldescriptors
 - cascading sorting functionality by allowing multiple sort query parameters on APIs
+- NEW API GET /api/assets/as-planned/distinctFilterValues
+- NEW API GET /api/assets/as-built/distinctFilterValues
 
 ### Changed
 - API BREAKING CHANGE: /api/assets changed to /api/assets/as-built